chore: structure changes + readme.md updated
