+++ app/core/exception_handler.py
@@ -10,7 +10,7 @@ from starlette.exceptions import HTTPException as StarletteHTTPException
 def register_error_handlers(app: FastAPI):
     # Validation error handler
     @app.exception_handler(RequestValidationError)
-    async def validation_exception_handler(request: Request, exc: RequestValidationError):
+    async def _validation_exception(request: Request, exc: RequestValidationError):
         response_content = BaseResponseDto(
             code="REQ_VALIDATION_FAILED",
             message="\n".join(
@@ -21,19 +21,11 @@ def register_error_handlers(app: FastAPI):
             )
         )
         return JSONResponse(status_code=422, content=response_content.model_dump())
-
-    # HTTPException handler
-    @app.exception_handler(HTTPException)
-    async def http_exception_handler(request: Request, exc: HTTPException):
-        response_content = BaseResponseDto(
-            code=exc.detail.code if isinstance(exc.detail, BaseResponseDto) else str(exc.status_code),
-            message=exc.detail.message if isinstance(exc.detail, BaseResponseDto) else str(exc.detail)
-        )
-        return JSONResponse(status_code=exc.status_code, content=response_content.model_dump(), headers=exc.headers)
     
-    # Starlette HTTPException handler (merged with above if no specific logic differs)
+    # Unified HTTPException handler for both FastAPI and Starlette exceptions
+    @app.exception_handler(HTTPException)
     @app.exception_handler(StarletteHTTPException)
-    async def starlette_http_exception_handler(request: Request, exc: StarletteHTTPException):
+    async def _http_exception(request: Request, exc: StarletteHTTPException):
         response_content = BaseResponseDto(
             code=exc.detail.code if isinstance(exc.detail, BaseResponseDto) else str(exc.status_code),
             message=exc.detail.message if isinstance(exc.detail, BaseResponseDto) else str(exc.detail)
@@ -42,6 +34,6 @@ def register_error_handlers(app: FastAPI):
 
     # Catch-all handler for unexpected errors
     @app.exception_handler(Exception)
-    async def global_exception_handler(request: Request, exc: Exception):
-        LOG.error(f"Unexpected error occurred: {exc}")
+    async def _global_exception(request: Request, exc: Exception):
+        LOG.error(f"Unexpected error occurred: {exc}", exc_info=True)
         return JSONResponse(status_code=500, content=InternalServerError().detail.model_dump())
