+++ app/server.py
@@ -6,8 +6,9 @@ from fastapi.middleware import Middleware
 from contextlib import asynccontextmanager
 from fastapi.middleware.cors import CORSMiddleware
 from app.res.openapi_error import GLOBAL_RESPONSES
+from app.middleware.handler import register_middlewares
 from app.core.exception_handler import register_error_handlers
-from app.middleware.middleware_handler import register_middlewares
+
